requirements.txt
SpeechRecognition
pyttsx3
pywhatkit
wikipedia
pyjokes
openai
python-dotenv
pyaudio